DRY + dead-code sweep: shared platform/MCP cores, canonical auth fallback
Implements the verified high-value findings from a four-agent parallel
audit of the whole tree (platform layer, operations, MCP/CLI/persistence,
tests/utils). Net effect: ~-830 lines.

Correctness fixes surfaced by the audit:
- quickQuery (and the plugin runner) resolve the claude binary via
  getClaudePath() like sessions do — a bare PATH lookup meant hosts with
  claude only at a common install location ran sessions fine while every
  haiku one-shot (routine/watch parses, watch confirms, distillation)
  silently failed.
- The Slack MCP platform API normalizes literal Unicode emoji to
  shortcodes (getEmojiName) for react_to_post and interactive-post
  reactions — reactions.add rejects raw emoji; the client path already
  normalized.
- The legacy sessionAllowedUsers fallback (a security invariant that was
  hand-copied at six call sites, and whose drift caused the recent
  approvals-owner lockout) lives once in authorization.ts as
  sessionAllowedUserSet().

Consolidations:
- base-client owns closeSocket() (the two clients' forceCloseConnection
  bodies were byte-identical); permalink types + the resolved-permalink
  formatter live once in permalink-shared.ts; the code-fence repair
  regex + rationale live once in platform/utils; dm-discovery-runtime
  uses dcmThreadId() instead of hand-building the prefix.
- mcp-server: nine tool registrations (each with its own eslint-disable)
  collapse into registerJsonTool(); the three post-list formatters share
  formatPostList(); the two limit clamps share clampLimit() (which
  clampThreadLimit now wraps).
- Tool formatters only set permissionText when it differs from display;
  the permission consumer falls back (12 verbatim echoes deleted).
- SystemExecutor's four post methods share one postLevel body;
  resolveMemoryConfig's inner bool() uses resolveBooleanFeature; the
  canonical SlackMcpApiConfig replaces the local shadow the factory cast
  was hiding.
- Tests: the byte-identical 96-line createMockSessionContext lives in
  test-utils (platform maker injected — the two files' platform mocks
  genuinely differ); both client test files adopt the shared fetch
  harness instead of verbatim local copies.

Dead code removed:
- tests/integration/helpers/assertions.ts (169 lines, zero refs),
  src/test-utils/wait-for.ts + test (zero importers), six dead exports
  in the integration wait-for helper, three dead session-helpers exports
  (two helpers with intra-file callers kept as private functions), and
  SlackClient.skipPermissions (assigned, never read).

Full battery run sequentially: typecheck, lint, knip, build,
bun test src/ (3139 pass), Slack integration suite (136 pass / 20
platform skips).
