+++ src/message-handler.ts
@@ -5,6 +5,7 @@
  * This ensures tests exercise the actual bot logic, not a duplicate.
  */
 
+import { sessionAllowedUserSet } from './session/authorization.js';
 import type { PlatformClient, PlatformPost, PlatformUser } from './platform/index.js';
 import type { SessionManager } from './session/index.js';
 import {
@@ -271,9 +272,7 @@ export async function handleMessage(
           // Clean up the paused session instead of resuming it
           const persistedSession = session.getPersistedSession(threadRoot);
           if (persistedSession) {
-            // Legacy records may lack sessionAllowedUsers — fall back to the
-            // owner (CLAUDE.md backward-compat rule; matches reaction-router).
-            const allowedUsers = new Set(persistedSession.sessionAllowedUsers || [persistedSession.startedBy].filter(Boolean));
+            const allowedUsers = sessionAllowedUserSet(persistedSession);
             if (allowedUsers.has(username) || client.isUserAllowed(username)) {
               auditLog(platformId, {
                 threadId: threadRoot,
@@ -300,11 +299,7 @@ export async function handleMessage(
       // reaction-router.ts — the platform allowlist alone is not enough.
       const persistedSession = session.getPersistedSession(threadRoot);
       if (persistedSession) {
-        // Legacy records may lack sessionAllowedUsers — without the owner
-        // fallback, `approvals: 'owner'` (which drops the global-allowlist
-        // rescue below) would lock the owner out of resuming their own
-        // session (CLAUDE.md backward-compat rule; matches reaction-router).
-        const allowedUsers = new Set(persistedSession.sessionAllowedUsers || [persistedSession.startedBy].filter(Boolean));
+        const allowedUsers = sessionAllowedUserSet(persistedSession);
         const ownerScoped =
           resolveApprovals(client.approvals, isDcmThreadId(threadRoot)) === 'owner';
         if (!allowedUsers.has(username) && (ownerScoped || !client.isUserAllowed(username))) {
