+++ src/platform/slack/client.ts
@@ -56,7 +56,6 @@ export class SlackClient extends BasePlatformClient {
   private botToken: string;
   private appToken: string;
   private channelId: string;
-  private skipPermissions: boolean;
   private apiUrl: string;
 
 
@@ -93,7 +92,6 @@ export class SlackClient extends BasePlatformClient {
     this.channelId = platformConfig.channelId;
     this.botName = platformConfig.botName;
     this.allowedUsers = platformConfig.allowedUsers;
-    this.skipPermissions = platformConfig.skipPermissions ?? false;
     this.apiUrl = platformConfig.apiUrl || 'https://slack.com/api';
     this.outboundFiles = platformConfig.outboundFiles;
     this.directChannelMode = resolveDirectChannelMode(platformConfig.directChannelMode);
@@ -581,32 +579,7 @@ export class SlackClient extends BasePlatformClient {
   protected forceCloseConnection(): Promise<void> {
     const ws = this.ws;
     this.ws = null;
-    if (!ws) return Promise.resolve();
-
-    ws.onopen = null;
-    ws.onmessage = null;
-    ws.onerror = null;
-
-    if (ws.readyState === WebSocket.CLOSED) {
-      ws.onclose = null;
-      return Promise.resolve();
-    }
-
-    return new Promise<void>((resolve) => {
-      const done = () => {
-        ws.onclose = null;
-        resolve();
-      };
-      ws.onclose = done;
-      setTimeout(done, 1000);
-      if (ws.readyState === WebSocket.OPEN || ws.readyState === WebSocket.CONNECTING) {
-        try {
-          ws.close();
-        } catch {
-          done();
-        }
-      }
-    });
+    return this.closeSocket(ws);
   }
 
   /**
