+++ tests/integration/helpers/wait-for.ts
@@ -51,71 +51,6 @@ export async function waitFor<T>(
 }
 
 /**
- * Wait for a condition to NOT throw an error
- *
- * @param fn - Function that may throw
- * @param options - Wait options
- * @returns The value returned by the function
- */
-export async function waitForNoThrow<T>(
-  fn: () => T | Promise<T>,
-  options: WaitOptions = {},
-): Promise<T> {
-  const { timeout = 10000, interval = 100, description = 'operation' } = options;
-
-  const startTime = Date.now();
-  let lastError: Error | null = null;
-
-  while (Date.now() - startTime < timeout) {
-    try {
-      return await fn();
-    } catch (error) {
-      lastError = error as Error;
-    }
-
-    await sleep(interval);
-  }
-
-  throw new Error(
-    `Timeout waiting for ${description} (${timeout}ms). Last error: ${lastError?.message}`,
-  );
-}
-
-/**
- * Wait for an array to have at least N items
- */
-export async function waitForLength<T>(
-  getArray: () => T[] | Promise<T[]>,
-  minLength: number,
-  options: WaitOptions = {},
-): Promise<T[]> {
-  return waitFor(
-    async () => {
-      const arr = await getArray();
-      return arr.length >= minLength ? arr : null;
-    },
-    { ...options, description: options.description || `array length >= ${minLength}` },
-  );
-}
-
-/**
- * Wait for a specific value in an array
- */
-export async function waitForInArray<T>(
-  getArray: () => T[] | Promise<T[]>,
-  predicate: (item: T) => boolean,
-  options: WaitOptions = {},
-): Promise<T> {
-  return waitFor(
-    async () => {
-      const arr = await getArray();
-      return arr.find(predicate) || null;
-    },
-    { ...options, description: options.description || 'item in array' },
-  );
-}
-
-/**
  * Retry an operation with exponential backoff
  */
 export async function retry<T>(
@@ -157,75 +92,4 @@ export async function retry<T>(
   }
 
   throw lastError;
-}
-
-/**
- * Create a deferred promise that can be resolved/rejected externally
- */
-export function createDeferred<T>(): {
-  promise: Promise<T>;
-  resolve: (value: T) => void;
-  reject: (error: Error) => void;
-} {
-  let resolve!: (value: T) => void;
-  let reject!: (error: Error) => void;
-
-  const promise = new Promise<T>((res, rej) => {
-    resolve = res;
-    reject = rej;
-  });
-
-  return { promise, resolve, reject };
-}
-
-/**
- * Wait for an event emitter to emit a specific event
- */
-export function waitForEvent<T>(
-  emitter: { on: (event: string, handler: (data: T) => void) => void },
-  event: string,
-  options: WaitOptions = {},
-): Promise<T> {
-  const { timeout = 10000, description = `event "${event}"` } = options;
-
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      reject(new Error(`Timeout waiting for ${description} (${timeout}ms)`));
-    }, timeout);
-
-    emitter.on(event, (data: T) => {
-      clearTimeout(timer);
-      resolve(data);
-    });
-  });
-}
-
-/**
- * Wait for a WebSocket to connect
- */
-export function waitForWebSocketOpen(
-  ws: WebSocket,
-  options: WaitOptions = {},
-): Promise<void> {
-  const { timeout = 10000 } = options;
-
-  if (ws.readyState === WebSocket.OPEN) {
-    return Promise.resolve();
-  }
-
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      reject(new Error(`WebSocket connection timeout (${timeout}ms)`));
-    }, timeout);
-
-    ws.addEventListener('open', () => {
-      clearTimeout(timer);
-      resolve();
-    });
-
-    ws.addEventListener('error', (event) => {
-      clearTimeout(timer);
-      reject(new Error(`WebSocket error: ${event}`));
-    });
-  });
 }
